+++ R/kokudosuuchi-package.R
@@ -1,10 +1,13 @@
-#' Kokudo Suuchi Joho - Download Japanese National Land Numerical Information
-#'
-#' Use the Kokudo Suuchi Joho API
+#' Utilities for 'Kokudo Suuchi'
 #'
 #' @docType package
 #' @name kokudosuuchi
-#' @importFrom dplyr %>%
-#' @importFrom rlang .data
-#' @seealso \url{http://nlftp.mlit.go.jp/ksj/api/about_api.html}
+#' @import rlang
+#' @seealso <http://nlftp.mlit.go.jp/ksj/api/about_api.html>
+"_PACKAGE"
+
+# The following block is used by usethis to automatically manage
+# roxygen namespace tags. Modify with care!
+## usethis namespace: start
+## usethis namespace: end
 NULL