+++ R/rest_in_peace.R
@@ -1,0 +1,22 @@
+#' Rest in peace, 'Kokudo Suuchi' API...
+#'
+#' @rdname RIP_API
+#' @param ... Ignored.
+#'
+#' @keywords internal
+#' @export
+getKSJSummary <- function(...) {
+  .Defunct()
+}
+
+#' @rdname RIP_API
+#' @export
+getKSJURL <- function(...) {
+  .Defunct()
+}
+
+#' @rdname RIP_API
+#' @export
+choose_prefecture_code <- function(...) {
+  .Defunct()
+}